pyproject.toml
[project]
name = "consensusai"
version = "1.0.0"
description = "ConsensusAI - Multi-model AI deliberation and peer review platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
]
